// src/types.ts
export enum MessageTypes {
  STANDARD = 'STANDARD',
  SELF_DESTRUCT_TIMED = 'SELF_DESTRUCT_TIMED',
  LIMITED_VIEW_TIME = 'LIMITED_VIEW_TIME',
}

export enum LogTypes {
  USER_LOGIN = 'USER_LOGIN',
  USER_REGISTERED = 'USER_REGISTERED',
  USER_VERIFIED = 'USER_VERIFIED',
  USER_VERIFIED_FAILED = 'USER_VERIFIED_FAILED',
  STARTED_CONVERSATION = 'STARTED_CONVERSATION',
  DELETED_CONVERSATION = 'DELETED_CONVERSATION',
  RECOVERED_CONVERSATION = 'RECOVERED_CONVERSATION',
  UNLOCKED_CONVERSATION = 'UNLOCKED_CONVERSATION',
  UNLOCKED_CONVERSATION_ATTEMP_FAIL = 'UNLOCKED_CONVERSATION_ATTEMP_FAIL',
  SENT_STANDARD_MESSAGE = 'SENT_STANDARD_MESSAGE',
  SENT_SELF_DESTRUCT_TIMED_MESSAGE = 'SENT_SELF_DESTRUCT_TIMED_MESSAGE',
  SENT_LIMITED_VIEW_TIME_MESSAGE = 'SENT_LIMITED_VIEW_TIME_MESSAGE',
}